Moved check_cube_not_float64, find_percentile_coord and find_threshold_coord from cube_checker into metadata modules.  Updated imports and unit tests.
