+++ lib/improver/convection.py
@@ -34,8 +34,8 @@ import iris
 import numpy as np
 
 from improver.nbhood.nbhood import NeighbourhoodProcessing
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.threshold import BasicThreshold
-from improver.utilities.cube_checker import find_threshold_coordinate
 from improver.utilities.spatial import DifferenceBetweenAdjacentGridSquares
