+++ lib/improver/ensemble_copula_coupling/ensemble_copula_coupling.py
@@ -47,8 +47,9 @@ from improver.ensemble_copula_coupling.ensemble_copula_coupling_utilities \
             get_bounds_of_distribution,
             insert_lower_and_upper_endpoint_to_1d_array,
             restore_non_probabilistic_dimensions)
+from improver.metadata.probabilistic import (
+    find_percentile_coordinate, find_threshold_coordinate)
 from improver.utilities.cube_checker import (
-    find_percentile_coordinate, find_threshold_coordinate,
     check_for_x_and_y_axes, check_cube_coordinates)
 from improver.utilities.cube_manipulation import (
     concatenate_cubes, enforce_coordinate_ordering)
