+++ lib/improver/nowcasting/lightning.py
@@ -36,8 +36,8 @@ import numpy as np
 from iris.exceptions import ConstraintMismatchError
 
 from improver.nbhood.nbhood import NeighbourhoodProcessing
-from improver.utilities.cube_checker import (
-    check_cube_coordinates, find_threshold_coordinate)
+from improver.metadata.probabilistic import find_threshold_coordinate
+from improver.utilities.cube_checker import check_cube_coordinates
 from improver.utilities.rescale import apply_double_scaling, rescale
 from improver.utilities.temporal import (
     extract_nearest_time_point, iris_time_to_datetime)
