+++ lib/improver/tests/blending/spatial_weights/test_SpatiallyVaryingWeightsFromMask.py
@@ -43,8 +43,8 @@ from iris.tests import IrisTest
 from iris.util import squeeze
 
 from improver.blending.spatial_weights import SpatiallyVaryingWeightsFromMask
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.tests.set_up_test_cubes import set_up_probability_cube
-from improver.utilities.cube_checker import find_threshold_coordinate
 from improver.utilities.warnings_handler import ManageWarnings
