+++ lib/improver/tests/ensemble_copula_coupling/ensemble_copula_coupling/test_GeneratePercentilesFromProbabilities.py
@@ -43,13 +43,13 @@ from iris.tests import IrisTest
 
 from improver.ensemble_copula_coupling.ensemble_copula_coupling import (
     GeneratePercentilesFromProbabilities as Plugin)
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.tests.ensemble_calibration.ensemble_calibration.\
     helper_functions import (
         add_forecast_reference_time_and_forecast_period,
         set_up_probability_threshold_cube,
         set_up_probability_above_threshold_temperature_cube,
         set_up_probability_above_threshold_spot_temperature_cube)
-from improver.utilities.cube_checker import find_threshold_coordinate
 from improver.utilities.warnings_handler import ManageWarnings
