+++ lib/improver/tests/metadata/test_enforce_datatypes_units.py
@@ -43,6 +43,102 @@ from improver.tests.set_up_test_cubes import (
     set_up_variable_cube, set_up_probability_cube, set_up_percentile_cube)
 
 
+class Test_check_cube_not_float64(IrisTest):
+
+    """Test whether a cube contains any float64 values."""
+
+    def setUp(self):
+        """Set up a cube to test."""
+        self.cube = set_up_variable_cube(np.ones((5, 5), dtype=np.float32),
+                                         spatial_grid='equalarea')
+
+    def test_float32_ok(self):
+        """Test a cube that should pass."""
+        enforce.check_cube_not_float64(self.cube)
+
+    def test_float64_cube_data(self):
+        """Test a failure of a cube with 64 bit data."""
+        self.cube.data = self.cube.data.astype(np.float64)
+        msg = "64 bit cube not allowed"
+        with self.assertRaisesRegex(TypeError, msg):
+            enforce.check_cube_not_float64(self.cube)
+
+    def test_float64_cube_data_with_fix(self):
+        """Test a cube with 64 bit data is converted to 32 bit data."""
+        self.cube.data = self.cube.data.astype(np.float64)
+        expected_cube = self.cube.copy()
+        expected_cube.data = expected_cube.data.astype(np.float64)
+        enforce.check_cube_not_float64(self.cube, fix=True)
+        self.assertEqual(self.cube, expected_cube)
+
+    def test_float64_cube_coord_points(self):
+        """Test a failure of a cube with 64 bit coord points."""
+        self.cube.coord("projection_x_coordinate").points = (
+            self.cube.coord("projection_x_coordinate").points.astype(
+                np.float64)
+        )
+        msg = "64 bit coord points not allowed"
+        with self.assertRaisesRegex(TypeError, msg):
+            enforce.check_cube_not_float64(self.cube)
+
+    def test_float64_cube_coord_points_with_fix(self):
+        """Test a failure of a cube with 64 bit coord points."""
+        self.cube.coord("projection_x_coordinate").points = (
+            self.cube.coord("projection_x_coordinate").points.astype(
+                np.float64))
+        expected_cube = self.cube.copy()
+        expected_cube.coord("projection_x_coordinate").points = (
+            expected_cube.coord("projection_x_coordinate").points.astype(
+                np.float64))
+        expected_coord = expected_cube.coord("projection_x_coordinate")
+        enforce.check_cube_not_float64(self.cube, fix=True)
+        self.assertEqual(self.cube, expected_cube)
+        self.assertEqual(
+            self.cube.coord("projection_x_coordinate"), expected_coord)
+
+    def test_float64_cube_coord_bounds(self):
+        """Test a failure of a cube with 64 bit coord bounds."""
+        x_coord = self.cube.coord("projection_x_coordinate")
+        # Default np.array for float input is np.float64.
+        x_coord.bounds = (
+            np.array([(point - 10., point + 10.) for point in x_coord.points])
+        )
+        msg = "64 bit coord bounds not allowed"
+        with self.assertRaisesRegex(TypeError, msg):
+            enforce.check_cube_not_float64(self.cube)
+
+    def test_float64_cube_coord_bounds_with_fix(self):
+        """Test a failure of a cube with 64 bit coord bounds."""
+        x_coord = self.cube.coord("projection_x_coordinate")
+        # Default np.array for float input is np.float64.
+        x_coord.bounds = (
+            np.array([(point - 10., point + 10.) for point in x_coord.points])
+        )
+        expected_cube = self.cube.copy()
+        expected_cube.coord("projection_x_coordinate").points = (
+            expected_cube.coord("projection_x_coordinate").points.astype(
+                np.float64))
+        expected_coord = expected_cube.coord("projection_x_coordinate")
+        enforce.check_cube_not_float64(self.cube, fix=True)
+        self.assertEqual(self.cube, expected_cube)
+        self.assertEqual(
+            self.cube.coord("projection_x_coordinate"), expected_coord)
+
+    def test_float64_cube_time_coord_points_ok(self):
+        """Test a pass of a cube with 64 bit time coord points."""
+        self.cube.coord("time").points = (
+             self.cube.coord("time").points.astype(np.float64))
+        enforce.check_cube_not_float64(self.cube)
+
+    def test_float64_cube_forecast_ref_time_coord_points_ok(self):
+        """Test a pass of a cube with 64 bit fcast ref time coord points."""
+        frt_coord = iris.coords.AuxCoord(
+            [np.float64(min(self.cube.coord("time").points))],
+            standard_name="forecast_reference_time")
+        self.cube.add_aux_coord(frt_coord)
+        enforce.check_cube_not_float64(self.cube)
+
+
 class Test_enforce_units_and_dtypes(IrisTest):
     """Test checking with option of enforcement or failure"""
