+++ lib/improver/tests/nowcasting/lightning/test_NowcastLightning.py
@@ -42,10 +42,10 @@ from iris.tests import IrisTest
 from iris.util import squeeze
 
 from improver.nowcasting.lightning import NowcastLightning as Plugin
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.tests.set_up_test_cubes import (
     set_up_variable_cube, set_up_probability_cube)
-from improver.utilities.cube_checker import (
-    find_dimension_coordinate_mismatch, find_threshold_coordinate)
+from improver.utilities.cube_checker import find_dimension_coordinate_mismatch
 
 
 def set_up_lightning_test_cubes(validity_time=dt(2015, 11, 23, 7),
