+++ lib/improver/tests/utilities/test_ProbabilitiesFromPercentiles2D.py
@@ -40,7 +40,7 @@ from iris.coords import DimCoord
 from iris.exceptions import CoordinateNotFoundError
 from iris.tests import IrisTest
 
-from improver.utilities.cube_checker import find_percentile_coordinate
+from improver.metadata.probabilistic import find_percentile_coordinate
 from improver.utilities.cube_manipulation import build_coordinate
 from improver.utilities.statistical_operations import \
     ProbabilitiesFromPercentiles2D
