+++ lib/improver/tests/utilities/test_cube_constraints.py
@@ -35,9 +35,9 @@ import unittest
 import iris
 from iris.tests import IrisTest
 
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.tests.utilities.test_cube_extraction import (
     set_up_precip_probability_cube)
-from improver.utilities.cube_checker import find_threshold_coordinate
 from improver.utilities.cube_constraints import create_sorted_lambda_constraint
