+++ lib/improver/tests/utilities/test_cube_extraction.py
@@ -38,11 +38,10 @@ import numpy as np
 from iris.exceptions import CoordinateNotFoundError
 from iris.tests import IrisTest
 
-from improver.utilities.cube_checker import find_threshold_coordinate
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.utilities.cube_extraction import (
     create_range_constraint, apply_extraction, extract_subcube,
     is_complex_parsing_required, create_constraint, parse_constraint_list)
-
 from improver.tests.set_up_test_cubes import set_up_probability_cube
