+++ lib/improver/tests/utilities/test_load.py
@@ -38,10 +38,10 @@ import iris
 import numpy as np
 from iris.tests import IrisTest
 
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.tests.set_up_test_cubes import (
     set_up_variable_cube, set_up_percentile_cube, set_up_probability_cube,
     add_coordinate)
-from improver.utilities.cube_checker import find_threshold_coordinate
 from improver.utilities.load import load_cube, load_cubelist
 from improver.utilities.save import save_netcdf
