+++ lib/improver/tests/wxcode/wxcode/test_WeatherSymbols.py
@@ -38,9 +38,9 @@ from cf_units import Unit
 from iris.coords import AuxCoord
 from iris.tests import IrisTest
 
+from improver.metadata.probabilistic import find_threshold_coordinate
 from improver.tests.ensemble_calibration.ensemble_calibration. \
     helper_functions import set_up_probability_threshold_cube
-from improver.utilities.cube_checker import find_threshold_coordinate
 from improver.wxcode.weather_symbols import WeatherSymbols
 from improver.wxcode.wxcode_utilities import WX_DICT
