+++ lib/improver/utilities/cube_manipulation.py
@@ -39,8 +39,8 @@ import six
 from iris.coords import AuxCoord, DimCoord
 from iris.exceptions import CoordinateNotFoundError
 
-from improver.utilities.cube_checker import (
-    check_cube_coordinates, find_threshold_coordinate)
+from improver.metadata.probabilistic import find_threshold_coordinate
+from improver.utilities.cube_checker import check_cube_coordinates
 
 
 def equalise_cube_attributes(cubes, silent=None):
