+++ lib/improver/wind_calculations/wind_direction.py
@@ -35,8 +35,8 @@ import numpy as np
 from iris.coords import CellMethod
 
 from improver.nbhood.nbhood import NeighbourhoodProcessing
-from improver.utilities.cube_checker import (
-    check_cube_coordinates, check_cube_not_float64)
+from improver.metadata.enforce_datatypes_units import check_cube_not_float64
+from improver.utilities.cube_checker import check_cube_coordinates
 
 
 class WindDirection(object):
